+++ samples/versions-in-platform/app/src/main/java/module-info.java
@@ -1,6 +1,7 @@
 module org.my.app {
-    requires org.slf4j;
     requires org.my.lib;
+    requires org.slf4j;
+
     requires static org.apache.xmlbeans;
 
     exports org.my.app;